pyproject.toml
[tool.poetry]
name = "nalog-lk"
version = "0.1.0"
description = ""
authors = ["Ildar Minnakhmetov <ildarworld@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
pytz = "^2021.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.7b0"
pytest-dotenv = "^0.5.2"
pytest-env = "^0.6.2"
pytest-asyncio = "^0.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
